[build] move GLFW-related files to platform/glfw
