+++ cmake/core-files.cmake
@@ -138,11 +138,9 @@ set(MBGL_CORE_FILES
     src/mbgl/platform/log.cpp
 
     # platform/default
-    include/mbgl/platform/default/glfw_view.hpp
     include/mbgl/platform/default/headless_backend.hpp
     include/mbgl/platform/default/headless_display.hpp
     include/mbgl/platform/default/offscreen_view.hpp
-    include/mbgl/platform/default/settings_json.hpp
     include/mbgl/platform/default/thread_pool.hpp
 
     # programs
